// src/model/store/language-code.ts
export enum LanguageCode {
  EN = 'en',
  CS = 'cs',
  DA = 'da',
  ET = 'et',
  ES = 'es',
  FI = 'fi',
  HU = 'hu',
  IT = 'it',
  LV = 'lv',
  LT = 'lt',
  NL = 'nl',
  NB = 'nb',
  PL = 'pl',
  PT = 'pt',
  RO = 'ro',
  RU = 'ru',
  DE = 'de',
  FR = 'fr',
  SK = 'sk',
  SV = 'sv',
  KO = 'ko',
  ZH = 'zh',
  JA = 'ja',
  TR = 'tr'
}